reparé le menu pause (la ref du bouton a sauté jsp pq) + réparé le roi qui atteint la derniere case du camp ennemi + tableau de int fait et donc id du player 2 est devenu -1

+++ Assets/Scripts/Piece/Piece.cs
@@ -9,8 +9,8 @@ public class Piece : MonoBehaviour,IPawn
 {
     public SOPiece soPiece;
     public Player player;
-    public short idPlayer;
-    public short idPiece;
+    public int idPlayer;
+    public int idPiece;
 
     private void Start()
     {
